requirements.txt
-i https://pypi.org/simple
bcrypt==4.0.1
black==23.1.0
cffi==1.15.1
cfgv==3.3.1 ; python_full_version >= '3.6.1'
click==8.1.3 ; python_version >= '3.7'
cryptography[ssh]==39.0.2
distlib==0.3.6
filelock==3.9.0 ; python_version >= '3.7'
flake8==6.0.0
flask==2.2.3
identify==2.5.19 ; python_version >= '3.7'
isort==5.12.0
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
mccabe==0.7.0 ; python_version >= '3.6'
mypy-extensions==1.0.0 ; python_version >= '3.5'
nodeenv==1.7.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
packaging==23.0 ; python_version >= '3.7'
pathspec==0.11.0 ; python_version >= '3.7'
platformdirs==3.1.0 ; python_version >= '3.7'
pre-commit==3.1.1
psycopg==3.1.8
psycopg2==2.9.5
pycodestyle==2.10.0 ; python_version >= '3.6'
pycparser==2.21
pyflakes==3.0.1 ; python_version >= '3.6'
pyjwt[crypto]==2.6.0
python-dotenv==1.0.0
pyyaml==6.0 ; python_version >= '3.6'
setuptools==67.6.0 ; python_version >= '3.7'
sqlalchemy==2.0.5.post1
tomli==2.0.1 ; python_version < '3.11'
typing-extensions==4.5.0 ; python_version >= '3.7'
virtualenv==20.20.0 ; python_version >= '3.7'
werkzeug==2.2.3 ; python_version >= '3.7'
